+++ Changelog.md
@@ -4,6 +4,9 @@
 
 ### SDFormat 9.1.0 (201X-XX-XX)
 
+1. Parser: add boolean parameter to `readFile` and `readString` for disabling conversion to latest SDF spec.
+    * [Pull request 647](https://bitbucket.org/osrf/sdformat/pull-requests/647)
+
 1. Added accessors to `ignition::math::[Boxd, Cylinderd, Planed, Sphered]`
    in the matching `sdf::[Box, Cylinder, Plane, Sphere]` classes.
     * [Pull request 639](https://bitbucket.org/osrf/sdformat/pull-requests/639)